+++ functions.c
@@ -5,12 +5,12 @@
 #include <math.h>
 
 struct process {
-    int arr_time;               // make unsigned long int
-    int pid;                    // make unsigned long int
-    int exec_time;
+    unsigned long int arr_time;               // make unsigned long int
+    unsigned long int pid;                    // make unsigned long int
+    unsigned long int exec_time;
     char *parallelisability;    // need to conver to just character later
-    int rem_exec_time;
-    int completed_time;
+    unsigned long int rem_exec_time;
+    unsigned long int completed_time;
     struct cpu *cpu_ptr;        // Should contain the reference/pointer to the CPU
     struct process *next;       // Pointer pointing to the next process stored in the Linked List
 
@@ -21,7 +21,7 @@ struct cpu {
     int cpu_id;
     // struct process* processes[PROCESSES_ARR_LENGTH]; // NEED to be an array of pointer to processes
     struct process_address_container* process_address_container_head;
-    int cpu_rem_exec_time;
+    unsigned long int cpu_rem_exec_time;
     struct process* running_process_ptr; // Pointer to the process that the CPU is currently running
 };
 
@@ -74,7 +74,7 @@ struct process* set_cpu_running_process_ptr(struct cpu* cpu_ptr)
     struct process_address_container* temp =  cpu_ptr->process_address_container_head;
     struct process* min_ptr = NULL;
 
-    int min = INT_MAX;
+    unsigned long min = ULONG_MAX;
     // Check loop while head not equal to NULL
     while (temp != NULL) {
         if (temp->process_ptr->rem_exec_time != 0 ) { // SKIP THE PROCESS THAT HAS FINISHED
@@ -108,7 +108,7 @@ void add_process_to_cpu(struct process* process, int num_cpus, struct cpu **cpu_
     // If tie between remainig times, then check CPU ids of tied ones
     // Assign to the one with smallest cpu_id
     
-    int min = INT_MAX;
+    unsigned long int min = ULONG_MAX;
     // int cpu_time;
     struct cpu* min_ptr = NULL;
 
@@ -179,7 +179,7 @@ struct process* get_process_with_smallest_rem_time_accounting_for_duplicates(str
     // it with UINT_MAX value.
     // UINT_MAX is integer type and its value
     // is 32767 or greater.
-    int min = INT_MAX;
+    unsigned long int min = ULONG_MAX;
   
     // Check loop while head not equal to NULL
     while (temp != NULL) {
@@ -215,7 +215,7 @@ struct process* get_process_with_smallest_rem_time_breaking_ties(struct process*
     // it with UINT_MAX value.
     // UINT_MAX is integer type and its value
     // is 32767 or greater.
-    int min = INT_MAX;
+    unsigned long int min = ULONG_MAX;
   
     // Check loop while head not equal to NULL
     while (temp != NULL) {
@@ -312,7 +312,7 @@ bool has_same_arrival_times(struct process* head)
     return false;
 }
 
-bool search(struct process* head, int x)
+bool search(struct process* head, unsigned long int x)
 {
     struct process* current = head;  // Initialize current
     while (current != NULL)
@@ -325,7 +325,7 @@ bool search(struct process* head, int x)
     return false;
 }
 
-struct process* get_pointer_to_process_equal_to_curr_time(struct process* head, int time)
+struct process* get_pointer_to_process_equal_to_curr_time(struct process* head, unsigned long int time)
 {
     struct process* current = head;  // Initialize current
     while (current != NULL)
@@ -358,7 +358,7 @@ bool is_all_process_completed(struct process* head){
 // smallest element
 struct process* get_shortest_rem_exec_time_process(struct process* head)
 {
-    int first = 0, second = INT_MAX; // Here first is equal to 0 is because, we know the first lowest rem_exec_time is 0 as one process already finished
+    unsigned long int first = 0, second = ULONG_MAX; // Here first is equal to 0 is because, we know the first lowest rem_exec_time is 0 as one process already finished
   
     struct process* temp = head;
     struct process* ptr;
@@ -380,7 +380,7 @@ struct process* get_shortest_rem_exec_time_process(struct process* head)
         temp = temp->next;
     }
   
-    if (second == INT_MAX){
+    if (second == ULONG_MAX){
         // cout << "There is no second smallest element\n";
         return NULL;
     }
@@ -399,7 +399,7 @@ struct process* smallestArrivalTime(struct process* head)       // NEED TO SORT 
     // it with UINT_MAX value.
     // UINT_MAX is integer type and its value
     // is 32767 or greater.
-    int min = INT_MAX;
+    unsigned long int min = ULONG_MAX;
   
     // Check loop while head not equal to NULL
     while (temp != NULL) {
@@ -418,7 +418,7 @@ struct process* smallestArrivalTime(struct process* head)       // NEED TO SORT 
 void printList(struct process* p)
 {
     while (p != NULL) {
-        printf("\n%d\n", p->arr_time);
+        printf("\n%lu\n", p->arr_time);
         p = p->next;
     }
 }
@@ -426,11 +426,14 @@ void printList(struct process* p)
 void print_CPU_process_list(struct process_address_container* p)
 {
     while (p != NULL) {
-        printf(" %d ", p->process_ptr->pid);
+        printf(" %lu ", p->process_ptr->pid);
         p = p->next;
     }
     printf("\n");
 }
+
+
+
 
 
 // void add_all_processes_arriving_at_same_time(struct process* head, int num_cpus, struct cpu **cpu_array, int current_time){
@@ -446,167 +449,6 @@ void print_CPU_process_list(struct process_address_container* p)
 //         }
 //     }
 // }
-
-
-    // struct process* find_lowest_arr_time(struct process* head_ptr) {
-
-    //     struct process* temp = head_ptr;
-    //     struct process* lowest;
-
-    //     while (temp->next != NULL){
-
-            
-            
-    //         if ((temp->arr_time) > (temp->next->arr_time)){
-    //             lowest = temp->next;
-    //         }
-    //         // else if ((temp->arr_time) == (temp->next->arr_time)){
-    //         //     lowest = lowest_pid();
-    //         // }
-    //         if ((temp->arr_time) < (temp->next->arr_time)){
-    //             lowest = temp;
-    //         }
-    //         temp = temp->next; //increment
-
-    //         // if (temp->next == NULL) {
-    //         //     return lowest;
-    //         // }
-    //     }
-    //     return lowest;
-
-    // }
-
-
-    // int calc_remaining_cpu_exec_time(int cpu_id, struct cpu **cpu_array){
-
-//     int cpu_arr_length = sizeof(*cpu_array)/sizeof(struct cpu);
-//     int cpu_rem_exec_time;
-//     for (int i=0; i<cpu_arr_length; i++){
-//         if (cpu_id == (*cpu_array)[i].cpu_id){  // If cpu with this id is found
-
-//             if ((*cpu_array)[i].cpu_id==-1){
-                
-//                 int processes_arr_length = sizeof((*cpu_array)[i].processes)/sizeof(struct process*);
-//                 for (int j=0; j<processes_arr_length; j++){ // for every process pointer in processes array
-//                     if((*cpu_array)[i].processes[j] == NULL){    //if process pointer is equal to NULL, i.e. "if no process is allocated yet" 
-//                         break; // break to assign cpu_rem_exec_time=0
-//                     } 
-//                 }
-                
-//                 cpu_rem_exec_time=0;// if cpu_id == -1, then cpu_rem_exec_time=0
-//             }
-//             else {// else, loop through each process pointer in the cpu's processes array and calculate total rem_exec_time of cpu
-//                 int cpu_rem_exec_time=0;
-
-//                 if ((*cpu_array)[i]) // CHECKS IF THERE ARE ANY PROCESSES IN THE PROCESS LIST OF THE CPU
-//                 int processes_arr_length = sizeof((*cpu_array)[i].processes)/sizeof(struct process*);
-//                 // printf("%d\n", processes_arr_length);
-//                 for (int j=0; j<processes_arr_length; j++){ // for every process pointer in processes array
-//                 cpu_rem_exec_time = cpu_rem_exec_time + (*cpu_array)[i].processes[j]->rem_exec_time; // access the process's rem_exec_time and add to cpu's rem_exec_time
-//                 }
-//             }
-//             return cpu_rem_exec_time;
-//         }
-//         else{
-//             printf("\nNO cpu with cpu_id %d\n", cpu_id);
-//         }
-//     }
-    
-//     return 0; // JUST FOR THE SAKE OF STOPPING WARNINGS
-// }
-
-// //SHOULD ADD THE PROCESS TO THE CPU
-// // AND ALSO, ADD THE CPU TO THE PROCESS
-// void add_process_to_cpu(struct process* process, struct cpu **cpu_array) {
-//     // Check which CPU's remaining execution time is lowest, and assign to that CPU
-//     // If tie between remainig times, then check CPU ids of tied ones
-//     // Assign to the one with smallest cpu_id
-
-//     int cpu_arr_length = sizeof(*cpu_array)/sizeof(struct cpu); // it can also be argv[x]
-//     int min = INT_MAX;
-//     int cpu_time;
-//     struct cpu* min_ptr;
-//     for (int i=0; i<cpu_arr_length; i++){
-
-//         cpu_time = calc_remaining_cpu_exec_time((*cpu_array)[i].cpu_id, cpu_array);
-
-//         if (cpu_time < min){
-//             min = cpu_time;
-//             min_ptr = &((*cpu_array)[i]); // Get the pointer to the CPU
-//         }
-//         else if (cpu_time == min){ // RESOLVE TIES WITH SAME CPU REMAINING EXECUTION TIME (This case will not happen until the value for the first CPU is calculated)
-//             // CHECK POINTER OF PREVIOUS CPU i.e. using i-1
-//             // CHECK FOR CPU WITH LOWEST CPU ID
-//             if ((*cpu_array)[i].cpu_id > (*cpu_array)[i-1].cpu_id){
-//                 //Put process in cpu with smaller cpu_id, i.e. i-1
-//                 min_ptr = &((*cpu_array)[i-1]);
-//             }
-//             else{
-//                 min_ptr = &((*cpu_array)[i]);
-//             }
-//         }
-//         else if (cpu_time > min){
-//             // Do Nothing - because you do not add it to the cpu with higher cpu_time than the one with lower one
-//         }
-
-//     }
-
-//     process->cpu_ptr = min_ptr;     //Add the cpu pointer to the process
-
-//     int proc_arr_len = sizeof(min_ptr->processes)/sizeof(struct process*);
-//     // struct processes** processes_new = (struct process**)realloc(min_ptr->processes, sizeof(struct process*)*(proc_arr_len + 1));
-//     min_ptr->processes = realloc(min_ptr->processes, sizeof(struct process*)*(proc_arr_len + 1));
-//     min_ptr->processes[proc_arr_len] = process;         // Add the process pointer to the processes array in the CPU
-    
-//     return;
-// }
-
-//void add_cpu_to_process(struct cpu* cpu, struct process *process) {
-    // Check which CPU's remaining execution time is lowest, and assign to that CPU
-    // If tie between remainig times, then check CPU ids of tied ones
-    // Assign to the one with smallest cpu_id
-
-
-//}
-
-// int calc_remaining_cpu_exec_time(int cpu_id, struct cpu **cpu_array){
-
-//     int cpu_arr_length = sizeof(*cpu_array)/sizeof(struct cpu);
-//     int cpu_rem_exec_time;
-//     for (int i=0; i<cpu_arr_length; i++){
-//         if (cpu_id == (*cpu_array)[i].cpu_id){  // If cpu with this id is found
-
-//             if ((*cpu_array)[i].cpu_id==-1){
-                
-//                 int processes_arr_length = sizeof((*cpu_array)[i].processes)/sizeof(struct process*);
-//                 for (int j=0; j<processes_arr_length; j++){ // for every process pointer in processes array
-//                     if((*cpu_array)[i].processes[j] == NULL){    //if process pointer is equal to NULL, i.e. "if no process is allocated yet" 
-//                         break; // break to assign cpu_rem_exec_time=0
-//                     } 
-//                 }
-                
-//                 cpu_rem_exec_time=0;// if cpu_id == -1, then cpu_rem_exec_time=0
-//             }
-//             else {// else, loop through each process pointer in the cpu's processes array and calculate total rem_exec_time of cpu
-//                 int cpu_rem_exec_time=0;
-
-//                 if ((*cpu_array)[i]) // CHECKS IF THERE ARE ANY PROCESSES IN THE PROCESS LIST OF THE CPU
-//                 int processes_arr_length = sizeof((*cpu_array)[i].processes)/sizeof(struct process*);
-//                 // printf("%d\n", processes_arr_length);
-//                 for (int j=0; j<processes_arr_length; j++){ // for every process pointer in processes array
-//                 cpu_rem_exec_time = cpu_rem_exec_time + (*cpu_array)[i].processes[j]->rem_exec_time; // access the process's rem_exec_time and add to cpu's rem_exec_time
-//                 }
-//             }
-//             return cpu_rem_exec_time;
-//         }
-//         else{
-//             printf("\nNO cpu with cpu_id %d\n", cpu_id);
-//         }
-//     }
-    
-//     return 0; // JUST FOR THE SAKE OF STOPPING WARNINGS
-// }
-
 
 
 // // Function that returns a linked list of processes with same arrival times
